refactoring done, all functionality restored, continuing development for bonus points

+++ app/tasks/models/Solution.java
@@ -15,6 +15,6 @@ public class Solution {
 
     private Long taskId;
 
-    private long[] items;
+    private int[] items;
 
 }
